# api/openapi.json
{"openapi":"3.1.0","info":{"title":"Synergy Matchmaker","description":"Synergy Matchmaker API","version":"0.4.0"},"paths":{"/users/signin":{"post":{"tags":["users"],"summary":"Signin","description":":return: User","operationId":"signin_users_signin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInEndpoint"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/User"},{"type":"null"}],"title":"Response Signin Users Signin Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/password":{"post":{"tags":["users"],"summary":"Update Password","description":":return: User","operationId":"update_password_users_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordEndpoint"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"boolean","title":"Response Update Password Users Password Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/renew_password":{"post":{"tags":["users"],"summary":"Update Password","description":":return: str","operationId":"update_password_users_renew_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenewPasswordEndpoint"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/user":{"patch":{"tags":["users"],"summary":"Update","description":":return: User","operationId":"update_users_user_patch","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["users"],"summary":"Delete","description":":return: bool","operationId":"delete_users_user_delete","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"boolean","title":"Response Delete Users User Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/match/":{"post":{"tags":["match"],"summary":"Match","operationId":"match_match__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/systems/test_version":{"get":{"tags":["system"],"summary":"Get Test Version","description":":return: List[Dict]\n[\n  {\n    \"id\": 1,\n    \"is_applied\": true,\n    \"tstamp\": \"2024-06-29T04:22:26.976814\",\n    \"version_id\": 0\n  },\n  {\n    \"id\": 2,\n    \"is_applied\": true,\n    \"tstamp\": \"2024-06-29T04:22:28.354794\",\n    \"version_id\": 1\n  }\n]","operationId":"get_test_version_systems_test_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Constraint":{"properties":{"max_num_teams":{"type":"integer","title":"Max Num Teams"},"members_per_team":{"type":"integer","title":"Members Per Team"},"at_least_one_pair_sex":{"type":"boolean","title":"At Least One Pair Sex","default":true},"girl_geq_boy":{"type":"boolean","title":"Girl Geq Boy","default":false},"boy_geq_girl":{"type":"boolean","title":"Boy Geq Girl","default":false},"at_least_one_leader":{"type":"boolean","title":"At Least One Leader","default":false},"unique_previous":{"type":"boolean","title":"Unique Previous","default":true},"group_diff_coeff":{"type":"number","title":"Group Diff Coeff","default":1.5}},"type":"object","required":["max_num_teams"],"title":"Constraint"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MatchingRequest":{"properties":{"student_constraints":{"items":{"$ref":"#/components/schemas/StudentConstraint"},"type":"array","title":"Student Constraints"},"constraint":{"$ref":"#/components/schemas/Constraint"}},"type":"object","required":["student_constraints","constraint"],"title":"MatchingRequest"},"RenewPasswordEndpoint":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"RenewPasswordEndpoint"},"SignInEndpoint":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"SignInEndpoint"},"StudentConstraint":{"properties":{"student_no":{"type":"integer","title":"Student No"},"dislikes":{"items":{"type":"integer"},"type":"array","title":"Dislikes","default":[]},"previous":{"type":"integer","title":"Previous"},"mi_a":{"type":"integer","title":"Mi A"},"mi_b":{"type":"integer","title":"Mi B"},"mi_c":{"type":"integer","title":"Mi C"},"mi_d":{"type":"integer","title":"Mi D"},"mi_e":{"type":"integer","title":"Mi E"},"mi_f":{"type":"integer","title":"Mi F"},"mi_g":{"type":"integer","title":"Mi G"},"mi_h":{"type":"integer","title":"Mi H"},"leader":{"type":"integer","title":"Leader"},"eyesight":{"type":"integer","title":"Eyesight"},"sex":{"type":"integer","title":"Sex"}},"type":"object","required":["student_no","previous","mi_a","mi_b","mi_c","mi_d","mi_e","mi_f","mi_g","mi_h","leader","eyesight","sex"],"title":"StudentConstraint"},"UpdatePasswordEndpoint":{"properties":{"user_id":{"type":"integer","title":"User Id"},"old_password":{"type":"string","title":"Old Password"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["user_id","old_password","new_password"],"title":"UpdatePasswordEndpoint"},"User":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"family_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family Name"},"given_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Given Name"},"school_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"School Id"},"email":{"type":"string","title":"Email"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"status":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Status"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"expired_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expired At"}},"type":"object","required":["id","email"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}